削除ロジックをMusicTebleコンポーネント内に内包して共通化。
